pyproject.toml
[tool.poetry]
name = "receipt-data-extraction"
version = "0.1.0"
description = ""
authors = ["ishan-modi <ishan.modi24@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.9"
tqdm = "^4.66.4"
jellyfish = "^1.0.3"
pillow = "^10.3.0"
mlflow = "^2.13.0"
boto3 = "^1.34.113"
transformers = "^4.41.1"
datasets = "^2.19.1"
torch = "^2.3.0"


[tool.poetry.group.dev.dependencies]
opencv-python = "^4.9.0.80"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
